servicedisco: fix a bug where service registrations would be lost
This PR fixes a bug where service registrations can be lost when registering
services of the same name into the same group (but with different tags). The
implementation of MakeAllocServiceID does not take tags into account when
creating what should be unique IDs for such services. This helper is used for
both Nomad and Consul service registrations.

For example,

group "group" {
    service {
      name     = "db"
      tags     = ["one"]
      provider = "nomad"
    }

    service {
      name     = "db"
      tags     = ["two"]
      provider = "nomad"
    }
}

In this case only one service would remain registered - as the second one in
the list would upsert and thus overwrite the first, because they share the
same ID generated by that helper function.

The fix is to take Tags into account in the helper function.