allocrunner: refactor coordinator and controller
The initial implementation wasn't very clear as to how all pieces were
connected together. This was the result of bad naming and encapsulation.
This commits fixes this by moving the new structs into their own package
since they are isolated from the rest of the allocrunner.

Moving them into their own package allows for simpler names that don't
have to repeat the task* prefix all the time and for better interfaces,
where methods that are expected to be called by external components are
now public and internal methods remain private. The package also has a
doc.go file with more extensive documentation.

The taskCoordinatorController struct was also poorly named (controller
is too generic) so it was renamed to simply Gate as it better reflects
its controls and behaviour.