core: connect with client before marking down due to missed heartbeat
This PR makes the Server agent(*) attempt to make a connection with a
Client before marking it as NodeStatusDown due to a missed heartbeat.

Manually double-checking responsiveness of the Client before marking
it down paves over issues where the Client may still be fine, but
coordination between Client, the server it is reporting status to,
and the leader managing heartbeat timers is not lined up.

(*) Who _should_ be the leader, but might not be in the immediate
aftermath of something triggering an election.

Potentially resolved causes of erroneous down-ed nodes:

1. Stale leader w/ expiring timers marking nodes it should not
2. Overloaded leader not responding to incoming update RPCs before timers expire