E2E/UI: set `NOMAD_PROXY_ADDR` separately from `NOMAD_ADDR` (#14587)
The Nomad UI will be proxied by the proxy job, but the test setup also requires
making direct HTTP API requests that bypass the UI. For clusters with mTLS
configured (ex. nightly), we need to bind-mount the user's mTLS certs picked up
by the API client into the test runner.

Includes a new README that explains the topology of the test environments in
detail and provides some advice for running locally.