checks: don't validate task/check compatibility for Connect services
In Nomad 1.4.0 we plumbed-through a task name to all checks in order to support
native service discovery. For Connect tasks with a non-script/gRPC check, this
triggers a validation error that setting the task name on a check isn't
compatible with http/tpc checks. Normally this is correct behavior but for
Connect tasks we need to include the `task` field. Bypass this validation for
Connect services.